Revert "Use ushort for id"
This reverts commit 30d79b9ac964e660103089d601b1cb702b141491.

+++ Fika.Core/Coop/GameMode/HostGameController.cs
@@ -210,7 +210,7 @@ namespace Fika.Core.Coop.GameMode
                 }
             }
 
-            ushort netId = 1000;
+            int netId = 1000;
             CoopBot coopBot;
             if (Bots.ContainsKey(profile.Id))
             {
